Users are assigned a color based on their position.
The color appears in their sinewave plot and also in the frequence plot.

Colors are set in js/colors.js.

For now, we take the index in a flattened 2D array (C order) and mod the
value by the length of the number of distinct colors. Presently, there
are 20 colors defined by d3's category20(). So some users will have
duplicate colors.